requirements.txt
click
pillow
